refactor helper functions and config
